drf-poll-api/requirements.txt
django
djangorestframework
psycopg2
pillow
model_bakery
gunicorn
django-environ
